requirements.txt
asgiref==3.7.2
Django==5.0.4
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.3.0
pytz==2024.1
sqlparse==0.4.4
